maildir spec compliance
The current implementation of mailpail isn't compliant with the maildir
specification, in that it reaches into the "cur" directory to replace
existing mails. Due to how Bitbucket versioning on pull requests
operate, this happens fairly frequently, especially in fast moving
respoitories.

This changeset attempts to preserve compliance with the maildir
specification by storing the last activity processed for a pull request
and creating mails for newer updates.