fix: generate more unique file names
I recently modified mailpail so that it never touches files after being
added to the "new" directory. As there is no longer a needs for fast
lookups of files for a message, we can generate filenames that are more
unique.

This changeset removes the "key" concept of filenames, as they no longer
need to be represented.